requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
acl-anthology-py=0.4.3=pypi_0
agentscope=0.0.6a3=pypi_0
aiofile=3.9.0=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiopath=0.6.11=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.7.1=pypi_0
app-paths=0.0.7=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=5.0.1=pypi_0
asyncstdlib=3.13.0=pypi_0
attrs=23.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bidict=0.23.1=pypi_0
black=24.10.0=pypi_0
bleach=6.2.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.11.26=h06a4308_0
cachetools=5.5.0=pypi_0
caio=0.9.17=pypi_0
certifi=2024.12.14=pypi_0
cffi=1.17.1=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
dashscope=1.14.1=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.11=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.15=pypi_0
dirtyjson=1.0.8=pypi_0
diskcache=5.6.3=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
docopt=0.6.2=pypi_0
docstring-parser=0.16=pypi_0
et-xmlfile=2.0.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
faiss-gpu=1.7.2=pypi_0
fastjsonschema=2.21.1=pypi_0
filelock=3.13.1=pypi_0
flake8=7.1.1=pypi_0
flask=3.0.0=pypi_0
flask-cors=4.0.0=pypi_0
flask-socketio=5.3.6=pypi_0
flask-sqlalchemy=3.1.1=pypi_0
fonttools=4.55.3=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-ai-generativelanguage=0.6.10=pypi_0
google-api-core=2.24.0=pypi_0
google-api-python-client=2.156.0=pypi_0
google-auth=2.37.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.8.3=pypi_0
googleapis-common-protos=1.66.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.68.1=pypi_0
grpcio-status=1.68.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.27.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
inputimeout=1.0.4=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.30.0=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.3=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
json5=0.10.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.11.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
kiwisolver=1.4.8=pypi_0
latexcodec=2.0.1=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
litellm=1.55.7=pypi_0
llama-index=0.10.30=pypi_0
llama-index-agent-openai=0.2.9=pypi_0
llama-index-cli=0.1.13=pypi_0
llama-index-core=0.10.68.post1=pypi_0
llama-index-embeddings-openai=0.1.11=pypi_0
llama-index-indices-managed-llama-cloud=0.1.6=pypi_0
llama-index-legacy=0.9.48.post4=pypi_0
llama-index-llms-openai=0.1.31=pypi_0
llama-index-multi-modal-llms-openai=0.1.9=pypi_0
llama-index-program-openai=0.1.7=pypi_0
llama-index-question-gen-openai=0.1.3=pypi_0
llama-index-readers-file=0.1.33=pypi_0
llama-index-readers-llama-parse=0.1.6=pypi_0
llama-parse=0.4.9=pypi_0
llamaindex-py-client=0.1.19=pypi_0
loguru=0.6.0=pypi_0
lxml=4.9.4=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.23.2=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
mccabe=0.7.0=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=pypi_0
nltk=3.9.1=pypi_0
notebook=7.3.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.1.105=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
ollama=0.4.4=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.58.1=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h5eee18b_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pathspec=0.12.1=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py310h06a4308_0
platformdirs=3.11.0=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.48=pypi_0
propcache=0.2.1=pypi_0
proto-plus=1.25.0=pypi_0
protobuf=5.29.2=pypi_0
psutil=6.1.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py=1.11.0=pypi_0
pyalex=0.15.1=pypi_0
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycodestyle=2.12.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pyflakes=3.2.0=pypi_0
pygments=2.18.0=pypi_0
pyjwt=2.8.0=pypi_0
pymongo=4.10.1=pypi_0
pyparsing=3.2.0=pypi_0
pypdf=4.3.1=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-engineio=4.11.1=pypi_0
python-json-logger=3.2.1=pypi_0
python-slugify=8.0.4=pypi_0
python-socketio=5.12.0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
retry=0.9.2=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.9.4=pypi_0
rnc2rng=2.7.0=pypi_0
rpds-py=0.22.3=pypi_0
rply=0.7.8=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
seaborn=0.13.2=pypi_0
send2trash=1.8.3=pypi_0
sentence-transformers=3.3.1=pypi_0
setuptools=75.1.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
simple-websocket=1.1.0=pypi_0
six=1.17.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlalchemy=2.0.36=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
strenum=0.4.15=pypi_0
striprtf=0.0.26=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.5.0=pypi_0
terminado=0.18.1=pypi_0
texsoup=0.3.1=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.8.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
together=1.3.11=pypi_0
tokenizers=0.21.0=pypi_0
tomli=2.2.1=pypi_0
torch=2.5.1+cu121=pypi_0
torchaudio=2.5.1+cu121=pypi_0
torchvision=0.20.1+cu121=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.47.1=pypi_0
triton=3.1.0=pypi_0
typer=0.15.1=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.2=pypi_0
unidecode=1.3.8=pypi_0
unpackable=0.0.4=pypi_0
uri-template=1.3.0=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py310h06a4308_0
wrapt=1.17.0=pypi_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.18.3=pypi_0
zhipuai=2.1.5.20241204=pypi_0
zipp=3.21.0=pypi_0
zlib=1.2.13=h5eee18b_1
